Removed redundant ActionController interface and renamed DefaultActionController to ActionController.
